# docs/api-spec/ncsa-dashboard-openapi.yaml
openapi: 3.0.3
info:
  title: NCSA Dashboard Web API
  version: 1.0.0
  description: |
    สัญญา API หลักสำหรับ `ncsa-dashboard-web`

    หลักการสำคัญ:
    - canonical path ใช้ `/api/v1/...`
    - response มาตรฐานใช้ `data`, `meta`, `error`
    - authentication ใช้ `Bearer token`
    - external sharing API ใช้ namespace `/api/v1/external/...` และ `X-API-Key`
    - web client สามารถเก็บ token ใน cookie ชื่อ `token` ได้ตามพฤติกรรมของ UI ปัจจุบัน
    - analytics ทั้งหมดใช้ timezone เริ่มต้น `Asia/Bangkok`
servers:
  - url: https://api.example.com
    description: Production placeholder
  - url: http://localhost:8000
    description: Local backend
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/executive/dashboard:
    get:
      tags: [Executive]
      summary: ข้อมูล executive dashboard ทั้งหน้า
      operationId: getExecutiveDashboard
      parameters:
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
      responses:
        '200':
          description: executive dashboard payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutiveDashboardResponse'
...
  /api/v1/operations/dashboard:
    get:
      tags: [Operations]
      summary: ข้อมูล operations dashboard ทั้งหน้า
      operationId: getOperationsDashboard
      x-compat-aliases:
        - path: /dashboard
          method: GET
        - path: /incidentbyseverity
          method: GET
        - path: /attacktime
          method: GET
        - path: /intelligencesources
          method: GET
        - path: /threattype
          method: GET
        - path: /countriesbythreatassociation
          method: GET
        - path: /targetsectors
          method: GET
      parameters:
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
      responses:
        '200':
          description: operations dashboard payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationsDashboardResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    StartDate:
      name: start_date
      in: query
      schema:
        type: string
        format: date
    EndDate:
      name: end_date
      in: query
      schema:
        type: string
        format: date
...
  schemas:
    ApiMeta:
      type: object
      properties:
        request_id:
          type: string
        generated_at:
          type: string
          format: date-time
        timezone:
          type: string
          example: Asia/Bangkok
      required: [generated_at, timezone]
...
    ErrorObject:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
      required: [code, message]
...
    ResponseBase:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/ApiMeta'
        error:
          $ref: '#/components/schemas/ErrorObject'
          nullable: true
      required: [meta, error]
...
    TopListItem:
      type: object
      properties:
        key:
          type: string
        label:
          type: string
        value:
          type: number
        percentage:
          type: number
          nullable: true
        color:
          type: string
          nullable: true
      required: [key, label, value]
    SeveritySlice:
      type: object
      properties:
        key:
          type: string
        label:
          type: string
        color:
          type: string
        value:
          type: integer
        percentage:
          type: number
      required: [key, label, color, value, percentage]
    PrimarySector:
      type: object
      properties:
        name:
          type: string
        value:
          type: integer
      required: [name, value]
    ThreatLevelCard:
      type: object
      properties:
        date:
          type: string
          format: date
        level:
          type: string
        level_th:
          type: string
        score:
          type: number
        delta_percent:
          type: number
          nullable: true
        primary_sector:
          $ref: '#/components/schemas/PrimarySector'
      required: [date, level, level_th, score, primary_sector]
    ExposureSummary:
      type: object
      properties:
        total_threats:
          type: integer
        ioc_active:
          type: integer
        critical_active:
          type: integer
        high_active:
          type: integer
      required: [total_threats, ioc_active, critical_active, high_active]
    TreemapNode:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        severity:
          type: string
        value:
          type: integer
      required: [id, label, severity, value]
    TreemapPayload:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/TreemapNode'
      required: [nodes]
    TrendPoint:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        total:
          type: integer
        critical:
          type: integer
        high:
          type: integer
        point_type:
          type: string
          enum: [historical, forecast]
      required: [timestamp, total, critical, high, point_type]
    ForecastTrendPayload:
      type: object
      properties:
        points:
          type: array
          items:
            $ref: '#/components/schemas/TrendPoint'
        forecast_start_index:
          type: integer
      required: [points, forecast_start_index]
    MapOrigin:
      type: object
      properties:
        country_code:
          type: string
          nullable: true
        country_name:
          type: string
        value:
          type: integer
        latitude:
          type: number
        longitude:
          type: number
      required: [country_name, value, latitude, longitude]
    MapConnection:
      type: object
      properties:
        origin_country:
          type: string
        target_country:
          type: string
        count:
          type: integer
      required: [origin_country, target_country, count]
    AttackOriginMap:
      type: object
      properties:
        target_country:
          type: string
        origins:
          type: array
          items:
            $ref: '#/components/schemas/MapOrigin'
        connections:
          type: array
          items:
            $ref: '#/components/schemas/MapConnection'
      required: [target_country, origins, connections]
    ExecutiveDashboardPayload:
      type: object
      properties:
        threat_level:
          $ref: '#/components/schemas/ThreatLevelCard'
        exposure_today:
          $ref: '#/components/schemas/ExposureSummary'
        severity_distribution:
          type: array
          items:
            $ref: '#/components/schemas/SeveritySlice'
        threat_volume_severity:
          $ref: '#/components/schemas/TreemapPayload'
        attack_volume_trend:
          $ref: '#/components/schemas/ForecastTrendPayload'
        attack_origin_map:
          $ref: '#/components/schemas/AttackOriginMap'
        filters:
          type: object
          additionalProperties: true
      required:
        - threat_level
        - exposure_today
        - severity_distribution
        - threat_volume_severity
        - attack_volume_trend
        - attack_origin_map
    OperationsOverview:
      type: object
      properties:
        active_ioc:
          type: integer
        critical_ioc_active:
          type: integer
        new_ioc:
          type: integer
        sources_active:
          type: integer
      required: [active_ioc, critical_ioc_active, new_ioc, sources_active]
    HeatmapCell:
      type: object
      properties:
        x:
          type: string
        y:
          type: string
        value:
          type: integer
      required: [x, y, value]
    HeatmapMatrix:
      type: object
      properties:
        mode:
          type: string
          example: day-hour
        x_axis:
          type: array
          items:
            type: string
        y_axis:
          type: array
          items:
            type: string
        cells:
          type: array
          items:
            $ref: '#/components/schemas/HeatmapCell'
      required: [mode, x_axis, y_axis, cells]
    OperationsDashboardPayload:
      type: object
      properties:
        overview:
          $ref: '#/components/schemas/OperationsOverview'
        incident_by_severity:
          type: array
          items:
            $ref: '#/components/schemas/SeveritySlice'
        attack_time_heatmap:
          $ref: '#/components/schemas/HeatmapMatrix'
        top_intelligence_sources:
          type: array
          items:
            $ref: '#/components/schemas/TopListItem'
        top_threat_types:
          type: array
          items:
            $ref: '#/components/schemas/TopListItem'
        top_attack_origins:
          type: array
          items:
            $ref: '#/components/schemas/TopListItem'
        target_sectors:
          type: array
          items:
            $ref: '#/components/schemas/TopListItem'
      required:
        - overview
        - incident_by_severity
        - attack_time_heatmap
        - top_intelligence_sources
        - top_threat_types
        - top_attack_origins
        - target_sectors
...
    ExecutiveDashboardResponse:
      allOf:
        - $ref: '#/components/schemas/ResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ExecutiveDashboardPayload'
          required: [data]
    OperationsDashboardResponse:
      allOf:
        - $ref: '#/components/schemas/ResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/OperationsDashboardPayload'
          required: [data]
...